Prediction/validation output for test data

The assign.output function outputs the summary results and plots for prediction/validation for the test dataset.

assign.output(processed.data, mcmc.pos.mean.testData, trainingData, testData,
  trainingLabel, testLabel, geneList, adaptive_B = TRUE, adaptive_S = FALSE,
  mixture_beta = TRUE, outputDir)

Arguments

processed.data
The list object returned from the assign.preprocess function.

mcmc.pos.mean.testData
The list object returned from the assign.mcmc function. Notice that for prediction/validation in the test dataset, the Y argument in the assign.mcmc function should be set as the test dataset.

trainingData
The genomic measure matrix of training samples (i.g., gene expression matrix). The dimension of this matrix is probe number x sample number.

testData
The genomic measure matrix of test samples (i.g., gene expression matrix). The dimension of this matrix is probe number x sample number.

trainingLabel
The list linking the index of each training sample to a specific group it belongs to.

testLabel
The vector of the phenotypes/labels of the test samples.

geneList
The list that collects the signature genes of one/multiple pathways. Every component of this list contains the signature genes associated with one pathway.

adaptive_B
Logicals. If TRUE, the model adapts the baseline/background (B) of genomic measures for the test samples. The default is TRUE.

adaptive_S
Logicals. If TRUE, the model adapts the signatures (S) of genomic measures for the test samples. The default is FALSE.

mixture_beta
Logicals. If TRUE, elements of the pathway activation matrix are modeled by a spike-and-slab mixture distribution. The default is TRUE.

outputDir
The path to the directory to save the output files. The path needs to be quoted in double quotation marks.

Value

The assign.output returns one .csv file containing one/multiple pathway activity for each individual test samples, scatter plots of pathway activity for each individual pathway in all the test samples, and heatmap plots for the gene expression of the prior signature and posterior signatures (if adaptive_S equals TRUE) of each individual pathway in the test samples.

Details

The assign.output function is suggested to run after the assign.preprocess, assign.mcmc and assign.summary functions. For the prediction/validation in the test dataset, The Y argument in the assign.mcmc function is the output value "testData_sub" from the assign.preprocess function.

Examples

assign.output(processed.data=processed.data,
              mcmc.pos.mean.testData=mcmc.pos.mean,
              trainingData=trainingData1, testData=testData1,
              trainingLabel=trainingLabel1, testLabel=testLabel1,
              geneList=NULL, adaptive_B=TRUE, adaptive_S=FALSE,
              mixture_beta=TRUE, outputDir=tempdir)

#> Outputing results...